requirements.txt
pydantic >= 1.4
python-dotenv
